requirements.txt
asgiref==3.2.10
Django==3.0.8
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
psycopg2-binary==2.8.5
Werkzeug==1.0.1
dnspython==2.0.0